Reorganised functions in controller into mapper to help with testing and to clean up the controller
