+++ src/EA.Weee.Web.Tests.Unit/Areas/Aatf/Controller/HomeControllerTests.cs
@@ -50,7 +50,12 @@
         public async void IndexGet_GivenValidViewModel_BreadcrumbShouldBeSet()
         {
             var organisationName = "Organisation";
+            var model = new HomeViewModel()
+            {
+                AatfList = A.Fake<List<AatfData>>()
+            };
 
+            A.CallTo(() => mapper.Map(A<AatfDataToHomeViewModelMapTransfer>._)).Returns(model);
             A.CallTo(() => cache.FetchOrganisationName(A<Guid>._)).Returns(organisationName);
 
             await controller.Index(A.Dummy<Guid>(), false);
@@ -74,6 +79,13 @@
         [Fact]
         public async void IndexGet_GivenAction_DefaultViewShouldBeReturned()
         {
+            var model = new HomeViewModel()
+            {
+                AatfList = A.Fake<List<AatfData>>()
+            };
+
+            A.CallTo(() => mapper.Map(A<AatfDataToHomeViewModelMapTransfer>._)).Returns(model);
+
             var result = await controller.Index(A.Dummy<Guid>(), false) as ViewResult;
 
             result.ViewName.Should().BeEmpty();
@@ -83,6 +95,12 @@
         public async void IndexGet_GivenOrganisationId_ApiShouldBeCalled()
         {
             var organisationId = Guid.NewGuid();
+            var model = new HomeViewModel()
+            {
+                AatfList = A.Fake<List<AatfData>>()
+            };
+
+            A.CallTo(() => mapper.Map(A<AatfDataToHomeViewModelMapTransfer>._)).Returns(model);
 
             await controller.Index(organisationId, false);
 
@@ -93,80 +111,22 @@
         public async void IndexGet_GivenOrganisationId_MapperShouldBeCalled()
         {
             var organisationId = Guid.NewGuid();
+            var model = new HomeViewModel()
+            {
+                AatfList = A.Fake<List<AatfData>>()
+            };
+
+            A.CallTo(() => mapper.Map(A<AatfDataToHomeViewModelMapTransfer>._)).Returns(model);
 
             await controller.Index(organisationId, false);
 
             A.CallTo(() => mapper.Map(A<AatfDataToHomeViewModelMapTransfer>.That.Matches(a => a.IsAE == false && a.OrganisationId == organisationId))).MustHaveHappened(Repeated.Exactly.Once);
         }
 
-        [Theory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public async void IndexGet_IsAEIsTrue_OnlyAEsShouldBeReturned(bool isAE)
-        {
-            var organisationId = Guid.NewGuid();
-            var aatfList = new List<AatfData>();
-
-            var aatfData = new AatfData(Guid.NewGuid(), "AATF", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
-                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
-                   A.Dummy<Core.Shared.PanAreaData>(), null)
-            {
-                FacilityType = FacilityType.Aatf
-            };
-
-            var aatfData2 = new AatfData(Guid.NewGuid(), "AATF2", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
-                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
-                   A.Dummy<Core.Shared.PanAreaData>(), null)
-            {
-                FacilityType = FacilityType.Aatf
-            };
-
-            var exporterData = new AatfData(Guid.NewGuid(), "AE", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
-                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
-                   A.Dummy<Core.Shared.PanAreaData>(), null)
-            {
-                FacilityType = FacilityType.Ae
-            };
-
-            var exporterData2 = new AatfData(Guid.NewGuid(), "AE", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
-               Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
-               A.Dummy<Core.Shared.PanAreaData>(), null)
-            {
-                FacilityType = FacilityType.Ae
-            };
-
-            aatfList.Add(aatfData);
-            aatfList.Add(exporterData);
-            aatfList.Add(aatfData2);
-            aatfList.Add(exporterData2);
-
-            A.CallTo(() => apiClient.SendAsync(A<string>._, A<GetAatfByOrganisation>.That.Matches(w => w.OrganisationId == organisationId))).Returns(aatfList);
-
-            var result = await controller.Index(organisationId, isAE) as ViewResult;
-
-            var model = result.Model as HomeViewModel;
-
-            if (isAE)
-            {
-                foreach (var aatf in model.AatfList)
-                {
-                    aatf.FacilityType.Should().Be(FacilityType.Ae);
-                }
-            }
-            else
-            {
-                foreach (var aatf in model.AatfList)
-                {
-                    aatf.FacilityType.Should().Be(FacilityType.Aatf);
-                }
-            }
-        }
-
         [Fact]
         public async void IndexGet_OnSingleAatf_PageRedirectsContactDetails()
         {
             var organisationId = Guid.NewGuid();
-
             var aatfList = new List<AatfData>();
 
             var aatfData = new AatfData(Guid.NewGuid(), "AATF", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
@@ -178,48 +138,19 @@
 
             aatfList.Add(aatfData);
 
+            var model = new HomeViewModel()
+            {
+                AatfList = aatfList
+            };
+
             A.CallTo(() => apiClient.SendAsync(A<string>._, A<GetAatfByOrganisation>._)).Returns(aatfList);
+            A.CallTo(() => mapper.Map(A<AatfDataToHomeViewModelMapTransfer>._)).Returns(model);
 
             var result = await controller.Index(organisationId, false) as RedirectToRouteResult;
 
             result.RouteValues["action"].Should().Be("Index");
             result.RouteValues["controller"].Should().Be("ViewAatfContactDetails");
             result.RouteValues["organisationId"].Should().Be(organisationId);
-        }
-
-        [Fact]
-        public async void IndexGet_HomeViewModelShouldBeBuilt()
-        {
-            var organisationId = Guid.NewGuid();
-
-            var aatfList = new List<AatfData>();
-
-            var aatfData = new AatfData(Guid.NewGuid(), "AATF", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
-                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
-                   A.Dummy<Core.Shared.PanAreaData>(), null)
-            {
-                FacilityType = FacilityType.Aatf
-            };
-
-            var aatfData2 = new AatfData(Guid.NewGuid(), "AATF2", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
-               Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
-               A.Dummy<Core.Shared.PanAreaData>(), null)
-            {
-                FacilityType = FacilityType.Aatf
-            };
-
-            aatfList.Add(aatfData);
-            aatfList.Add(aatfData2);
-
-            A.CallTo(() => apiClient.SendAsync(A<string>._, A<GetAatfByOrganisation>._)).Returns(aatfList);
-
-            var result = await controller.Index(organisationId, false) as ViewResult;
-
-            var model = result.Model as HomeViewModel;
-
-            model.IsAE.Should().Be(false);
-            model.AatfList.Should().BeEquivalentTo(aatfList);
-            model.OrganisationId.Should().Be(organisationId);
         }
 
         [Fact]
