+++ src/EA.Weee.Web.Tests.Unit/Areas/Aatf/Mapping/ToViewModel/AatfDataToHomeViewModelMapTests.cs
@@ -29,6 +29,92 @@
         }
 
         [Fact]
+        public void Map_GivenListOfAatfs_ContactDetailsNameIsBuilt()
+        {
+            var organisationId = Guid.NewGuid();
+            var aatfList = new List<AatfData>();
+
+            var aatfData = new AatfData(Guid.NewGuid(), "AATF", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
+                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
+                   A.Dummy<Core.Shared.PanAreaData>(), null)
+            {
+                FacilityType = FacilityType.Aatf
+            };
+
+            aatfList.Add(aatfData);
+
+            var transfer = new AatfDataToHomeViewModelMapTransfer() { OrganisationId = organisationId, AatfList = aatfList, IsAE = false };
+
+            var result = map.Map(transfer);
+
+            foreach (var aatf in result.AatfList)
+            {
+                aatf.AatfContactDetailsName.Should().Be(aatf.Name + " (" + aatf.ApprovalNumber + ") - " + aatf.AatfStatus);
+            }
+        }
+
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public void Map_GivenIsAE_AatfListShouldOnlyContainThatFacilityType(bool isAE)
+        {
+            var organisationId = Guid.NewGuid();
+            var aatfList = new List<AatfData>();
+
+            var aatfData = new AatfData(Guid.NewGuid(), "AATF", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
+                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
+                   A.Dummy<Core.Shared.PanAreaData>(), null)
+            {
+                FacilityType = FacilityType.Aatf
+            };
+
+            var aatfData2 = new AatfData(Guid.NewGuid(), "AATF2", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
+                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
+                   A.Dummy<Core.Shared.PanAreaData>(), null)
+            {
+                FacilityType = FacilityType.Aatf
+            };
+
+            var exporterData = new AatfData(Guid.NewGuid(), "AE", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
+                   Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
+                   A.Dummy<Core.Shared.PanAreaData>(), null)
+            {
+                FacilityType = FacilityType.Ae
+            };
+
+            var exporterData2 = new AatfData(Guid.NewGuid(), "AE", "approval number", 2019, A.Dummy<Core.Shared.UKCompetentAuthorityData>(),
+               Core.AatfReturn.AatfStatus.Approved, A.Dummy<AatfAddressData>(), Core.AatfReturn.AatfSize.Large, DateTime.Now,
+               A.Dummy<Core.Shared.PanAreaData>(), null)
+            {
+                FacilityType = FacilityType.Ae
+            };
+
+            aatfList.Add(aatfData);
+            aatfList.Add(exporterData);
+            aatfList.Add(aatfData2);
+            aatfList.Add(exporterData2);
+
+            var transfer = new AatfDataToHomeViewModelMapTransfer() { OrganisationId = organisationId, AatfList = aatfList, IsAE = isAE };
+
+            var result = map.Map(transfer);
+
+            if (isAE)
+            {
+                foreach (var aatf in result.AatfList)
+                {
+                    aatf.FacilityType.Should().Be(FacilityType.Ae);
+                }
+            }
+            else
+            {
+                foreach (var aatf in result.AatfList)
+                {
+                    aatf.FacilityType.Should().Be(FacilityType.Aatf);
+                }
+            }
+        }
+
+        [Fact]
         public void Map_GivenSource_PropertiesShouldBeSet()
         {
             var transfer = new AatfDataToHomeViewModelMapTransfer() { OrganisationId = Guid.NewGuid(), AatfList = A.Fake<List<AatfData>>(), IsAE = false };
