+++ src/EA.Weee.Web/Areas/Aatf/Mappings/ToViewModel/AatfDataToHomeViewModelMap.cs
@@ -2,7 +2,9 @@
 {
     using EA.Prsd.Core;
     using EA.Prsd.Core.Mapper;
+    using EA.Weee.Core.AatfReturn;
     using EA.Weee.Web.Areas.Aatf.ViewModels;
+    using System.Collections.Generic;
 
     public class AatfDataToHomeViewModelMap : IMap<AatfDataToHomeViewModelMapTransfer, HomeViewModel>
     {
@@ -16,9 +18,37 @@
         {
             Guard.ArgumentNotNull(() => source, source);
 
+            var selectedAatfsOrAes = new List<AatfData>();
+
+            if (source.IsAE)
+            {
+                foreach (var aatf in source.AatfList)
+                {
+                    if (aatf.FacilityType == Core.AatfReturn.FacilityType.Ae)
+                    {
+                        selectedAatfsOrAes.Add(aatf);
+                    }
+                }
+            }
+            else
+            {
+                foreach (var aatf in source.AatfList)
+                {
+                    if (aatf.FacilityType == Core.AatfReturn.FacilityType.Aatf)
+                    {
+                        selectedAatfsOrAes.Add(aatf);
+                    }
+                }
+            }
+
+            foreach (var aatf in selectedAatfsOrAes)
+            {
+                aatf.AatfContactDetailsName = aatf.Name + " (" + aatf.ApprovalNumber + ") - " + aatf.AatfStatus;
+            }
+
             Model.IsAE = source.IsAE;
             Model.OrganisationId = source.OrganisationId;
-            Model.AatfList = source.AatfList;
+            Model.AatfList = selectedAatfsOrAes;
 
             return Model;
         }
